refactor(iso): hoist serializeLocationForCache to its own module
It was duplicated between loader.tsx (exported) and use-loader-runner.tsx
(private copy to avoid circular import). One module owns it now.
loader.tsx re-exports so existing test imports keep working.
