+++ packages/iso/src/internal/loader.tsx
@@ -8,6 +8,7 @@ import type { LoaderRef } from '../define-loader.js';
 import { RouteLocationsContext } from './route-locations.js';
 import { ErrorBoundary } from './route-boundary.js';
 import { useLoaderRunner } from './use-loader-runner.js';
+export { serializeLocationForCache } from './cache-key.js';
 
 type LoaderProps<T> = {
   loader: LoaderRef<T>;
@@ -111,15 +112,3 @@ function DataReader<T>({
   );
 }
 
-export function serializeLocationForCache(
-  loc: RouteHook,
-  params: string[] | '*'
-): string {
-  const sp = (loc.searchParams ?? {}) as Record<string, string>;
-  const keys =
-    params === '*'
-      ? Object.keys(sp).sort()
-      : params.filter((k) => k in sp).sort();
-  const sortedSearch = keys.map((k) => `${k}=${sp[k]}`).join('&');
-  return `${loc.path}?${sortedSearch}`;
-}
